Separate business logic from serviceImpl to keep a clean architecture
